+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/solution/optimization/meta/Escalator.java
@@ -14,7 +14,7 @@ import net.splitcells.dem.data.set.list.List;
 import net.splitcells.dem.resource.communication.interaction.LogLevel;
 import net.splitcells.gel.rating.type.Cost;
 import net.splitcells.gel.solution.SolutionView;
-import net.splitcells.gel.solution.optimization.Optimization;
+import net.splitcells.gel.solution.optimization.OfflineOptimization;
 import net.splitcells.gel.solution.optimization.OptimizationEvent;
 
 import java.util.function.Function;
@@ -25,22 +25,22 @@ import static net.splitcells.dem.lang.perspective.PerspectiveI.perspective;
 import static net.splitcells.dem.resource.host.interaction.Domsole.domsole;
 import static net.splitcells.gel.common.Language.OPTIMIZATION;
 
-public class Escalator implements Optimization {
+public class Escalator implements OfflineOptimization {
 
-    public static Escalator escalator(Function<Integer, Optimization> optimizations) {
+    public static Escalator escalator(Function<Integer, OfflineOptimization> optimizations) {
         return new Escalator(optimizations, 0, 0, Integer.MAX_VALUE);
     }
 
-    public static Escalator escalator(Function<Integer, Optimization> optimizations, int escalationLevel, int minimum_escalation_level, int maximum_escalation_level) {
+    public static Escalator escalator(Function<Integer, OfflineOptimization> optimizations, int escalationLevel, int minimum_escalation_level, int maximum_escalation_level) {
         return new Escalator(optimizations, escalationLevel, minimum_escalation_level, maximum_escalation_level);
     }
 
-    private final Function<Integer, Optimization> optimizations;
+    private final Function<Integer, OfflineOptimization> optimizations;
     private int escalationLevel;
     private final int minimum_escalation_level;
     private final int maximum_escalation_level;
 
-    private Escalator(Function<Integer, Optimization> optimizations, int escalationLevel, int minimum_escalation_level, int maximum_escalation_level) {
+    private Escalator(Function<Integer, OfflineOptimization> optimizations, int escalationLevel, int minimum_escalation_level, int maximum_escalation_level) {
         this.optimizations = optimizations;
         this.escalationLevel = escalationLevel;
         this.minimum_escalation_level = minimum_escalation_level;